requirements.txt
# Core dependencies
Flask==3.1.0
flask-cors==4.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==21.2.0
pymongo==4.6.2
blinker==1.9.0

# Document processing
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
pillow==10.2.0
xlrd==2.0.1
openpyxl==3.1.2
pandas==2.2.2
tabula-py==2.9.0

# AI and ML
huggingface-hub>=0.23.0,<1.0
transformers==4.41.1
openai==1.14.1
mistralai==1.0.1
langchain==0.1.12
langchain-community==0.0.29
tiktoken==0.5.2
spacy==3.7.4
sentence-transformers==2.3.1

# Data handling
numpy==1.26.4
scipy==1.13.0
nltk==3.8.1
beautifulsoup4==4.12.3
lxml==5.2.0

# Testing
pytest==8.3.5
pytest-flask==1.3.0
pytest-cov==6.0.0

# Development tools
black==24.3.0
isort==5.12.0
flake8==7.0.0

# Document image processing
poppler-utils==0.1.0
Wand==0.6.13
opencv-python-headless # Added for CV operations
# Security
cryptography==42.0.8
PyJWT==2.8.0

# AWS integration (optional)
boto3==1.34.71
awscli==1.32.71
